# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MoneyMosaic API",
    "version": "1.0.0",
    "description": "A comprehensive personal finance dashboard API that connects multiple banks and tracks finances using the Plaid API.",
    "contact": {
      "name": "MoneyMosaic API Support",
      "url": "https://github.com/matteoguidii/moneymosaic"
    },
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "details": {
            "type": "string",
            "description": "Additional error details"
          }
        },
        "required": [
          "error"
        ]
      },
...
    }
  },
...
  "paths": {
...
    "/api/sandbox/public_token/create": {
      "post": {
        "summary": "Create sandbox public token",
        "description": "Creates a test public token for sandbox environment testing",
        "tags": [
          "Sandbox"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "institution_id": {
                    "type": "string",
                    "description": "Plaid institution ID for testing"
                  },
                  "initial_products": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Array of products to initialize"
                  }
                },
                "required": [
                  "institution_id",
                  "initial_products"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sandbox public token created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "public_token": {
                      "type": "string",
                      "description": "Test public token"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}